requirements.txt
psycopg2-binary
flask
flask_jwt_extended
python-dotenv
bcrypt
loguru
playwright
flask-swagger-ui